ragie-rag-agent/requirements.txt
ragie # Ragie SDK
python-dotenv>=1.0.0  # Environment variables management
fastapi   # Web framework
gunicorn>=20.1.0  # WSGI HTTP Server
openai>=1.68.2 # OpenAI API client
requests>=2.28.0  # HTTP requests
pydantic>=2.0.0  # Data validation
python-multipart>=0.0.6  # Form data handling
streamlit>=1.27.0  # Frontend UI framework 